# perf: early validate all messages before encoding