--- a/encoder/stream_test.go
+++ b/encoder/stream_test.go
@@ -146,6 +146,16 @@ func TestStreamEncoderUnhappyFlow(t *testing.T) {
 		t.Fatalf("expected err: %v, got: %v", io.EOF, err)
 	}
 
+	// Protocol validation error
+	streamEnc, _ = New(mockWriterAt{}).StreamEncoder()
+	streamEnc.enc.protocolValidator.ProtocolVersion = proto.V1
+	err = streamEnc.WriteMessage(&proto.Message{Fields: []proto.Field{
+		factory.CreateField(mesgnum.Record, fieldnum.RecordSpeed1S).WithValue(make([]uint8, 256)),
+	}, DeveloperFields: []proto.DeveloperField{{}}})
+	if !errors.Is(err, proto.ErrProtocolViolation) {
+		t.Fatalf("expected err: %v, got: %v", proto.ErrProtocolViolation, err)
+	}
+
 	// Message validation error
 	streamEnc, _ = New(mockWriterAt{}).StreamEncoder()
 	err = streamEnc.WriteMessage(&proto.Message{Fields: []proto.Field{