Validating messages early give several advantages:
1. If invalid message occurs, we can return early without doing unnecessary works such as marshaling, calculate CRC or even do a syscall to write to the file.
2. If normal writer is given (a writer without offset writer capability), the encoding speed is improved by geomean 16% since  we don't need to validate messages twices. Benchmark:
  ```go
  goos: linuxl goarch: amd64l pkg: benchfitl cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                             │   old.txt   │               new.txt               │
                             │   sec/op    │   sec/op     vs base                │
  Encode/muktihari/fit_raw-4   93.57m ± 8%   74.63m ± 3%  -20.24% (p=0.000 n=10)
  Encode/muktihari/fit-4       157.2m ± 7%   136.7m ± 2%  -13.01% (p=0.000 n=10)
  geomean                      121.3m        101.0m       -16.71%
  ```
3. We can release all objects generated by the message validator early, so it can be garbage-collected as soon as we are done validating. The objects are only be generated when it encounters developer data.